// types/next-auth.d.ts
import "next-auth";
import "next-auth/jwt";

export type Role = "owner" | "staff" | "agent" | "user";

declare module "next-auth" {
  interface Session {
    user: {
      id: string;
      name?: string | null;
      email?: string | null;
      role: Role;
    };
  }

  interface User {
    id: string;
    role: Role;
  }
}

declare module "next-auth/jwt" {
  interface JWT {
    id: string;
    role: Role;
  }
}